+++ SNAPLib/Read.cpp
@@ -40,7 +40,7 @@ Read::checkIdMatch(
 {
     if (!readIdsMatch(read0, read1)) {
         unsigned n[2] = {min(read0->getIdLength(), 200u), min(read1->getIdLength(), 200u)};
-        char* p[2] = {(char*) alloca(n[0] + 1), (char*) alloca(n[1] + 1)};
+        char* p[2] = {(char*) alloca((size_t)n[0] + 1), (char*) alloca((size_t)n[1] + 1)};
         memcpy(p[0], read0->getId(), n[0]); p[0][n[0]] = 0;
         memcpy(p[1], read1->getId(), n[1]); p[1][n[1]] = 0;
         WriteErrorMessage("Unmatched read IDs '%s' and '%s'.  Use the -I option to ignore this.\n", p[0], p[1]);
